+++ src/keyboard_input/input_sender_dummy.cpp
@@ -1,0 +1,6 @@
+#include "input_sender.h"
+
+void sendTokensAsInput( [[maybe_unused]] const std::vector<Token> tokens )
+{
+    // dummy
+}
