+++ src/media_keys/media_keys.cpp
@@ -1,0 +1,26 @@
+#include "media_keys.h"
+#include "src/keyboard_input/input_sender.h"
+
+namespace keyboardinput
+{
+void sendMediaNextSong()
+{
+    // noop
+}
+
+void sendMediaPreviousSong()
+{
+    // noop
+}
+
+void sendMediaPausePlay()
+{
+    // noop
+}
+
+void sendMediaStopSong()
+{
+    // noop
+}
+
+} // namespace keyboardinput
